+++ backend/canisters/local_user_index/api/src/updates/register_user.rs
@@ -6,6 +6,8 @@ use types::UserId;
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
     pub username: String,
+    #[serde(default)]
+    pub display_name: Option<String>,
     pub referral_code: Option<String>,
     pub public_key: Vec<u8>,
 }